working start

+++ engine/src/types.rs
@@ -4,7 +4,7 @@ use crate::core::Atomic;
 
 #[derive(Serialize, Deserialize, Copy, Clone, PartialEq, Default)]
 pub struct Location {
-    pub lat: f32,
-    pub lon: f32,
+    pub lat: f64,
+    pub lon: f64,
 }
 impl Atomic for Location {}
